Less alloc, mutable API instead of builder pattern (#40)
* WIP

* can run tests

* separate hydration into crate

* docs

* views can self replace with receivers

* patch children

* don't drop responders on view drop, clean up callbacks on view drop, no txrx::hand_clone + receiver clone is branch, don't store receivers

* replace view and patch view interface using receivers of Into<T>

* test patch children using Into<T>

* add new interfaces to builder + hydrator, update examples

* tests and readme

* ci pull requests

* wasm-pack build examples, cicd stuff

* include mdbook in tests

* mdbook uses its own build target

* mdbook deps

* Testing out a multipage build (#44)

* Create an example with multiple routes

Each route renders its content based only on the `Route`.

* Add logging and additional log levels to aid debugging/exploring

* Eliminate the `Out::SkipRender` message

There's no need for it, if we do not send a message with new route
information not attempt will be made to update the view. Thus
`SkipRender` is logically the equivalent of no message.

* Remove log statements

The log statements were a way to explore the sequence of events
surrounding element drops but are not useful for "app" operation.

* Use `From<Route>` to create a `View`

Rather than calling an `App::render` function. `render` sounds too much
like a pattern from other frameworks that does not match this mental
model.

* Return `ViewBuilder` from route functions

Use `From` to enable conversion from `Route` to `ViewBuilder` and from
`ViewBuilder` to `View`. The `Route` -> `ViewBuilder` is used for
initial render. `Route` -> `View` is used when a new `Route` is received
on the `Receiver`.

* separate coobook cicd and release

* separate cookbook cache key

* gizmos don't have views or builders

* cicd

* cicd

Co-authored-by: Bryan J Swift <mail@bryanwrit.es>