Add SSR to the multipage example (#51)
* Start to add additional complexity to the home route

* Add "game" components to display minesweeper (like) game board

* Add "route" for displaying minesweeper (like) game board

* Add MIT license to multipage example

* Add description and repository metadata to multipage example

* Change `KeyboardEvent` feature for `MouseEvent`

* Use `usize` rather than `u16` for coordinates

* Enable "bulk updates" to board view

* Add a way to fetch game state from a local server

* Fetch game state when starting to render the game

The game renders as a constant set of cells but update them once the
game loads.

* Change `board` function to take references

`Transmitter` and `Receiver` always have a `'static` lifetime (?) so can
be safely passed as references throughout the applicaiton.

* Add a `GameList` component to render list of game ids

Each `<li>` in the `GameList` (from `GameList::game_li`) should be
clickable and trigger a message which can be translated to a
`Route::Game`.

* Rename `GetGameError` to `FetchError`

* Add `api::get_game_list` to retrieve list of game ids

* Render the `GameList` component rather than a default `Game`

Change the "Games" link to navigate to a `GameList`.

* Use namespaces to clarify intent

* Use `<slot>` with `patch:children` to replace views

* Create `RouteDispatcher` to facilitate passing `Transmitter<Route>`

* Create a more general `fetch` function

Implement `get_game` and get_game_list` in terms of `fetch`.

* Add `patch_game` function to update remote game

* Enable playing the game

Clicking a cell triggers a PATCH to the server, the server responds with
an updated board state. The board view is replaced on each update.

* Use `Uuid` rater than `String` to represent ids

Using `Uuid` means `Copy` can be derived for structs and enums as well
because `Uuid` is `Copy`.

* Only include the web-sys features used

* Include `init` JS function from the server root

* Implement traits to translate Route <-> String

Use the `From` trait to turn a `String` into a `Route` and the `Display`
trait to turn a `Route` into a `String`.

* Change `RouteDispatcher::navigate` to `::push_state`

`RouteDispatcher::push_state` more closely resembles the underlying
browser based implementation which uses `History::push_state_with_url`
to add to the history stack.

* Load initial route from the browser location

Pass the initial route to the `App` setup.

* Make hydration work for the home page

* Only include `mogwai::utils` in the functions using it

* Create a `Route` from anything that can be treated as a `str` ref

* Only use the API during initial view if on wasm32

* Add a way to create SSR view of the application

* Add a "barebones" SSR implementation

* Remove the complexities of "Game" related functionality

* Remove unused features from `web-sys` dependency

* Remove unused crates from multipage example

* Add doc block for `tera::Tera` TEMPLATES definition

* Clone a `Transmitter` from a given reference in `RouteDispatcher::new`

* Move `push_state` to a `route_dispatch` module

Move `view_builder` function from `RouteDispatcher` to `From<Route>`
implementation for `ViewBuilder<HtmlElement>`. Remove the
`RouteDispatcher` struct.

* Do not create `Transmitter` and `Receiver` for `App`

Or rather, let `Gizmo::from` do it for us.