Change `Transmitter::send_async` only exist for wasm32 (#58)
* Change `Transmitter::send_async` to log a message for not wasm32

The `futures::executor::block_on` didn't work inside a `tokio` runtime
which will be too common a scenario to ignore. Outside of implementing a
custom `Waker` to loop and poll the `Future` to completion spawning a
"local" context similar to the wasm32 behavior is more difficult than
anticipated.

Refs #52

* Remove `futures` from the dependencies of mogwai

Without the usage in `Transmitter::send_async` the crate is not
necesary.

Closes #57

* Move the `send_async` example to `Transmitter::send_async`

To ensure the example is only compiled for `target_arch = "wasm32"`.